Remove rocketfuel code, keep a small compatibility layer for the API (bug 1106586, bug 1058292)
