+++ mkt/api/tests/nose.cfg
@@ -17,10 +17,6 @@ tests=mkt.abuse.tests.test_views,
       mkt.api.tests.test_urls,
       mkt.api.tests.test_views,
       mkt.comm.tests.test_views,
-      mkt.collections.tests.test_authorization,
-      mkt.collections.tests.test_fields,
-      mkt.collections.tests.test_serializers,
-      mkt.collections.tests.test_views,
       mkt.developers.tests.test_api_payments,
       mkt.developers.tests.test_views_api,
       mkt.features.tests.test_serializers,
