--- mkt/site/fixtures/data/collection_81721.json
@@ -1,60 +1,0 @@
-[
-{
-    "pk": 580, 
-    "model": "translations.translation", 
-    "fields": {
-        "localized_string_clean": "Public apps", 
-        "created": "2014-04-17T10:20:36", 
-        "locale": "en-us", 
-        "modified": "2014-04-17T10:20:36", 
-        "id": 1363, 
-        "localized_string": "Public apps"
-    }
-},
-{
-    "pk": 579, 
-    "model": "translations.translation", 
-    "fields": {
-        "localized_string_clean": "Public apps", 
-        "created": "2014-04-17T10:20:36", 
-        "locale": "en-us", 
-        "modified": "2014-04-17T10:20:36", 
-        "id": 1362, 
-        "localized_string": "Public apps"
-    }
-},
-{
-    "pk": 81721, 
-    "model": "collections.collection", 
-    "fields": {
-        "category": null, 
-        "curators": [], 
-        "image_hash": null, 
-        "description": 1362, 
-        "created": "2014-04-17T10:20:36", 
-        "region": null, 
-        "author": "Mark", 
-        "default_language": "en-US", 
-        "modified": "2014-04-17T10:20:49", 
-        "can_be_hero": false, 
-        "text_color": "#000000", 
-        "carrier": null, 
-        "collection_type": 0, 
-        "is_public": true, 
-        "background_color": "#ffffff", 
-        "slug": "public-apps", 
-        "name": 1363
-    }
-},
-{
-    "pk": 8, 
-    "model": "collections.collectionmembership", 
-    "fields": {
-        "app": 337141, 
-        "order": 0, 
-        "modified": "2014-04-17T10:20:44", 
-        "collection": 81721, 
-        "created": "2014-04-17T10:20:44"
-    }
-}
-]
